Introduce warnings if diffs
