+++ avaframe/com1DFA/standardTests_com1DFACfg.ini
@@ -1,0 +1,48 @@
+### Config File - This file contains the main settings for the simulation run
+## Copy to  local_com1DFACfg.ini and set you parameters
+# This file is part of Avaframe.
+
+
+[GENERAL]
+# Path to samos executable and AK_Attributes files
+samosAT = /home/anna/repos/gitea/AvaFrame_SamosDFA/samosAT -files /home/anna/repos/gitea/AvaFrame_SamosDFA/files/AK_Attributes/
+# Name of run cint file template
+RunCint = runBasic
+# Input format of release are files
+# -options: shp (shapfile), nxyz (list of polygon coordinates x,y,z=0))
+inputf = shp
+# True if full output shall be printed to terminal
+flagOut = False
+# True if entrainment areas shall be included
+flagEnt = True
+# True if resistance areas shall be included
+flagRes = True
+
+
+[PARAMETERVAR]
+# True if parameter variation of MU option shall be included
+flagVarPar = False
+# True if entres simulation shall be used for parameter variation
+flagVarEnt = True
+# parameter to be varied, current option: Mu (friction parameter)
+varPar = RelTh
+# list of parameter values, individual values need to be separated by '_'
+varParValues = 0.750
+# Name of run cint file template if parameter variation is used
+varRunCint = vary
+
+
+[DEFVALUES]
+# standard value for Mu
+Mu = 0.155
+# standard value for release area thickness
+RelTh = 1.000
+
+
+[REPORT]
+# which result parameters shall be included as plots in report
+plotFields = ppr_pv_pfd
+# units for output variables
+unitppr = kPa
+unitpfd = m
+unitpv = ms-1
